+++ source/pham/db/db_fbexception.d
@@ -21,13 +21,13 @@ class FbException : SkException
 public:
     this(string message, int code, string state,
         int socketCode = 0, int vendorCode = 0,
-        string file = __FILE__, size_t line = __LINE__, Throwable next = null) pure
+        string file = __FILE__, uint line = __LINE__, Throwable next = null) pure
     {
         super(message, code, state, socketCode, vendorCode, file, line, next);
     }
 
     this(FbIscStatues statues,
-        string file = __FILE__, size_t line = __LINE__, Throwable next = null)
+        string file = __FILE__, uint line = __LINE__, Throwable next = null)
     {
         string statusMessage, statusState;
         int statusCode;
