+++ source/pham/db/db_myexception.d
@@ -21,13 +21,13 @@ class MyException : SkException
 public:
     this(string message, int code, string sqlState,
         int socketCode = 0, int vendorCode = 0,
-        string file = __FILE__, size_t line = __LINE__, Throwable next = null) pure
+        string file = __FILE__, uint line = __LINE__, Throwable next = null) pure
     {
         super(message, code, sqlState, socketCode, vendorCode, file, line, next);
     }
 
     this(MyErrorResult errorResult,
-        string file = __FILE__, size_t line = __LINE__, Throwable next = null) pure
+        string file = __FILE__, uint line = __LINE__, Throwable next = null) pure
     {
         super(errorResult.message, errorResult.code, errorResult.sqlState, 0, errorResult.code, file, line, next);
         this.errorResult = errorResult;
