+++ source/pham/db/db_pgexception.d
@@ -22,13 +22,13 @@ class PgException : SkException
 public:
     this(string message, int code, string sqlState,
         int socketCode = 0, int vendorCode = 0,
-        string file = __FILE__, size_t line = __LINE__, Throwable next = null) pure
+        string file = __FILE__, uint line = __LINE__, Throwable next = null) pure
     {
         super(message, code, sqlState, socketCode, vendorCode, file, line, next);
     }
 
     this(PgGenericResponse statues,
-        string file = __FILE__, size_t line = __LINE__, Throwable next = null)
+        string file = __FILE__, uint line = __LINE__, Throwable next = null)
     {
         auto statusMessage = statues.errorString();
         auto statusSqlState = statues.sqlState();
